requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
aiohttp==3.9.1
aiohttp-jinja2==1.5.1
Jinja2==3.1.2
aiofiles==23.2.1
Pillow==10.1.0
requests==2.31.0